--- packages/components/modules/comments/native/CommentContainer/styles.ts
@@ -1,11 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-export const createStyles = () =>
-  StyleSheet.create({
-    contentContainer: {
-      display: 'flex',
-      flexDirection: 'column',
-      flex: 1,
-      paddingHorizontal: 16,
-    },
-  })
