+++ packages/components/modules/comments/native/Comments/styles.ts
@@ -1,6 +1,8 @@
+import { Theme } from '@baseapp-frontend/design-system/styles/native'
+
 import { StyleSheet } from 'react-native'
 
-export const createStyles = () =>
+export const createStyles = (theme: Theme) =>
   StyleSheet.create({
     transparent: {
       backgroundColor: 'transparent',
@@ -16,5 +18,25 @@ export const createStyles = () =>
       flexDirection: 'column',
       flex: 1,
       paddingHorizontal: 16,
+    },
+    bottomDrawerActionContainer: {
+      paddingHorizontal: 16,
+      paddingVertical: 12,
+      gap: 4,
+      borderBottomWidth: 1,
+      borderBottomColor: theme.colors.surface.disabled,
+    },
+    bottomDrawerDivider: {
+      paddingHorizontal: 16,
+      paddingVertical: 12,
+      gap: 4,
+      borderBottomWidth: 0,
+    },
+    bottomDrawerPressable: {
+      flexDirection: 'row',
+      alignItems: 'center',
+      justifyContent: 'flex-start',
+      gap: 8,
+      padding: 8,
     },
   })
