+++ packages/components/modules/comments/native/index.ts
@@ -1,8 +1,5 @@
 // exports native comments components
 
-export { default as CommentContainer } from './CommentContainer'
-export type * from './CommentContainer/types'
-
 export { default as CommentItem } from './CommentItem'
 export type * from './CommentItem/types'
